Modified homepage and channel
