+++ app/components/homePage.js
@@ -1,13 +1,20 @@
 'use strict'
 
 var React = require('react')
+
 // var Router = require('react-router')
 // var Link = Router.Link
 
 var Home = React.createClass({
   render: function () {
     return (
-      <div>
+      <div className='bodyContainer'>
+
+        <div className='contentContainer'>
+
+          <h3>Welcome to Cylinder</h3>
+          <p>Enjoy your <i className='fa fa-youtube'></i>YouTube in peace</p>
+          <p>Get started by creating your own channel or join an existing channel</p>
         <div id='container'>
           <div id='monitor'>
             <div id='monitorscreen'>
@@ -15,8 +22,9 @@ var Home = React.createClass({
             </div>
           </div>
         </div>
-
+      </div>
         <div className='channelsContainer'>
+
         </div>
     </div>
     )
